% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catalog_cache_table}
\alias{catalog_cache_table}
\alias{catalog_uncache_table}
\alias{catalog_clear_cache}
\alias{catalog_is_cached}
\title{Cache And Uncache Tables}
\usage{
catalog_cache_table(sc, table)

catalog_uncache_table(sc, table)

catalog_clear_cache(sc)

catalog_is_cached(sc, table)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{table}{\code{character(1)}. The name of the table.}
}
\value{
\itemize{
\item \code{catalog_cache_table()}: \code{NULL}, invisibly.
}

\itemize{
\item \code{catalog_uncache_table()}: \code{NULL}, invisibly.
}

\itemize{
\item \code{catalog_clear_cache()}: \code{NULL}, invisibly.
}

\itemize{
\item \code{catalog_is_cached()}: A \code{logical(1)} vector indicating \code{TRUE} if the table
is cached and \code{FALSE} otherwise.
}
}
\description{
Spark SQL can cache tables using an in-memory columnar format by calling
\code{catalog_cache_table(sc, "tableName")}. Spark SQL will scan only required
columns and will automatically tune compression to minimize memory usage and
GC pressure. You can call \code{catalog_uncache_table(sc, "tableName")} to remove
the table from memory. Similarly you can call \code{catalog_clear_cache(sc)} to
remove all cached tables from the in-memory cache.
}
\examples{
\dontrun{
sc <- sparklyr::spark_connect(master = "local")
mtcars_spark <- sparklyr::copy_to(dest = sc, df = mtcars)

# By default the table is not cached
catalog_is_cached(sc = sc, table = "mtcars")

# We can manually cache the table
catalog_cache_table(sc = sc, table = "mtcars")
# And now the table is cached
catalog_is_cached(sc = sc, table = "mtcars")

# We can uncache the table
catalog_uncache_table(sc = sc, table = "mtcars")
catalog_is_cached(sc = sc, table = "mtcars")
}

}
\seealso{
\code{\link[=catalog_create_table]{catalog_create_table()}}, \code{\link[=catalog_get_table]{catalog_get_table()}}, \code{\link[=catalog_list_tables]{catalog_list_tables()}},
\code{\link[=catalog_refresh_table]{catalog_refresh_table()}}, \code{\link[=catalog_table_exists]{catalog_table_exists()}},
\code{\link[=catalog_uncache_table]{catalog_uncache_table()}}
}
