% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catalog_create_table}
\alias{catalog_create_table}
\title{Create A Table}
\usage{
catalog_create_table(sc, table, path, source, ...)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{table}{\code{character(1)}. The name of the table to create.}

\item{path}{\code{character(1)}. The path to use to create the table.}

\item{source}{\code{character(1)}. The data source to use to create the table such
as \code{"parquet"}, \code{"csv"}, etc.}

\item{...}{Additional options to be passed to the \code{createTable} method.}
}
\description{
Creates a table, in the hive warehouse, from the given path and returns the
corresponding \code{DataFrame}. The table will contain the contents of the file
that is in the \code{path} parameter.
}
\details{
The default data source type is parquet.
This can be changed using \code{source} or setting the configuration option
\code{spark.sql.sources.default} when creating the spark session using or after
you have created the session using\if{html}{\out{<div class="r">}}\preformatted{config <- sparklyr::spark_config()
config[["spark.sql.sources.default"]] <- "csv"
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=catalog_cache_table]{catalog_cache_table()}}, \code{\link[=catalog_get_table]{catalog_get_table()}}, \code{\link[=catalog_list_tables]{catalog_list_tables()}},
\code{\link[=catalog_refresh_table]{catalog_refresh_table()}}, \code{\link[=catalog_table_exists]{catalog_table_exists()}},
\code{\link[=catalog_uncache_table]{catalog_uncache_table()}}
}
