% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{catalog_current_database}
\alias{catalog_current_database}
\title{Get The Current Database}
\usage{
catalog_current_database(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}.}
}
\value{
\code{character(1)}, the current database name.
}
\description{
Returns the current database in this session. By default your session will be
connected to the "default" database (named "default") and to change database
you can use \code{\link[=catalog_set_current_database]{catalog_set_current_database()}}.
}
\examples{
\dontrun{
sc <- sparklyr::spark_connect(master = "local")
catalog_current_database(sc = sc)
}

}
\seealso{
\code{\link[=catalog_set_current_database]{catalog_set_current_database()}}, \code{\link[=catalog_database_exists]{catalog_database_exists()}},
\code{\link[=catalog_list_databases]{catalog_list_databases()}}
}
