% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catalog_get_table}
\alias{catalog_get_table}
\title{Get A Table}
\usage{
catalog_get_table(sc, table, database = NULL)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{table}{\code{character(1)}. The name of the table.}

\item{database}{\code{character(1)}. The name of the database for which the
functions should be listed (default: \code{NULL}).}
}
\description{
Get the table or view with the specified name in the specified database. You
can use this to find the table's description, database, type and whether it
is a temporary table or not.
}
\seealso{
\code{\link[=catalog_cache_table]{catalog_cache_table()}}, \code{\link[=catalog_create_table]{catalog_create_table()}}, \code{\link[=catalog_list_tables]{catalog_list_tables()}},
\code{\link[=catalog_refresh_table]{catalog_refresh_table()}}, \code{\link[=catalog_table_exists]{catalog_table_exists()}},
\code{\link[=catalog_uncache_table]{catalog_uncache_table()}}
}
