% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catalog_refresh_table}
\alias{catalog_refresh_table}
\title{Refresh a Table}
\usage{
catalog_refresh_table(sc, table)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{table}{\code{character(1)}. The name of the table.}
}
\value{
\code{NULL}
}
\description{
Invalidates and refreshes all the cached data and metadata of the given
table. For performance reasons, Spark SQL or the external data source library
it uses might cache certain metadata about a table, such as the location of
blocks. When those change outside of Spark SQL, users should call this
function to invalidate the cache. If this table is cached as an
\code{InMemoryRelation}, drop the original cached version and make the new version
cached lazily.
}
\seealso{
\code{\link[=catalog_cache_table]{catalog_cache_table()}}, \code{\link[=catalog_create_table]{catalog_create_table()}}, \code{\link[=catalog_get_table]{catalog_get_table()}},
\code{\link[=catalog_list_tables]{catalog_list_tables()}}, \code{\link[=catalog_table_exists]{catalog_table_exists()}}, \code{\link[=catalog_uncache_table]{catalog_uncache_table()}}
}
