% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{catalog_set_current_database}
\alias{catalog_set_current_database}
\title{Set The Current Database}
\usage{
catalog_set_current_database(sc, name)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{name}{\code{character(1)}. The name of the database to set the current
database to.}
}
\description{
Sets the current default database in this session.
}
\examples{
\dontrun{
sc <- sparklyr::spark_connect(master = "local")
catalog_set_current_database(sc = sc, name = "new_db")
}

}
\seealso{
\code{\link[=catalog_current_database]{catalog_current_database()}}, \code{\link[=catalog_database_exists]{catalog_database_exists()}},
\code{\link[=catalog_list_databases]{catalog_list_databases()}}
}
