% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catalog_table_exists}
\alias{catalog_table_exists}
\title{Check If A Table Exists}
\usage{
catalog_table_exists(sc, table, database = NULL)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{table}{\code{character(1)}. The name of the table.}

\item{database}{\code{character(1)}. The name of the database for which the
functions should be listed (default: \code{NULL}).}
}
\value{
A \code{logical(1)} vector indicating \code{TRUE} if the table exists within the
specified database and \code{FALSE} otherwise.
}
\description{
Check if the table or view with the specified name exists in the specified
database. This can either be a temporary view or a table/view.
}
\details{
If \code{database} is \code{NULL}, \code{catalog_table_exists} refers to a table in the
current database (see \code{\link[=catalog_current_database]{catalog_current_database()}}).
}
\examples{
\dontrun{
sc <- sparklyr::spark_connect(master = "local")
mtcars_spark <- sparklyr::copy_to(dest = sc, df = mtcars)
catalog_table_exists(sc = sc, table = "mtcars")
}

}
\seealso{
\code{\link[=catalog_cache_table]{catalog_cache_table()}}, \code{\link[=catalog_create_table]{catalog_create_table()}}, \code{\link[=catalog_get_table]{catalog_get_table()}},
\code{\link[=catalog_list_tables]{catalog_list_tables()}}, \code{\link[=catalog_refresh_table]{catalog_refresh_table()}}, \code{\link[=catalog_uncache_table]{catalog_uncache_table()}}
}
