% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{current_database}
\alias{current_database}
\title{Get The Current Database}
\usage{
current_database(sc)
}
\arguments{
\item{sc}{A \code{spark_connection}.}
}
\value{
\code{character(1)}, the current database name.
}
\description{
Returns the current database in this session. By default your session will be
connected to the "default" database (named "default") and to change database
you can use \code{\link[=set_current_database]{set_current_database()}}.
}
\examples{
\dontrun{
sc <- sparklyr::spark_connect(master = "local")
current_database(sc = sc)
}

}
\seealso{
\code{\link[=set_current_database]{set_current_database()}}, \code{\link[=database_exists]{database_exists()}}, \code{\link[=list_databases]{list_databases()}}
}
