% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_table}
\alias{get_table}
\title{Get A Table}
\usage{
get_table(sc, table, database = NULL)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{table}{\code{character(1)}. The name of the table.}

\item{database}{\code{character(1)}. The name of the database for which the
functions should be listed (default: \code{NULL}).}
}
\value{
An object of class \code{spark_jobj} and \code{shell_jobj}.
}
\description{
Get the table or view with the specified name in the specified database. You
can use this to find the table's description, database, type and whether it
is a temporary table or not.
}
\seealso{
\code{\link[=cache_table]{cache_table()}}, \code{\link[=create_table]{create_table()}}, \code{\link[=list_tables]{list_tables()}}, \code{\link[=refresh_table]{refresh_table()}},
\code{\link[=table_exists]{table_exists()}}, \code{\link[=uncache_table]{uncache_table()}}
}
