% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_partitions.R
\name{recover_partitions}
\alias{recover_partitions}
\title{Recover Partitions}
\usage{
recover_partitions(sc, table)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{table}{\code{character(1)}. The name of the table.}
}
\description{
Recovers all the partitions in the directory of a table and update the
catalog. This only works for partitioned tables and not un-partitioned tables
or views.
}
