% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_by_path.R
\name{refresh_by_path}
\alias{refresh_by_path}
\title{Refresh By Path}
\usage{
refresh_by_path(sc, path)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{path}{\code{character(1)}. The path to refresh.}
}
\description{
Invalidates and refreshes all the cached data (and the associated metadata)
for any Dataset that contains the given data source path. Path matching is by
prefix, i.e. "/" would invalidate everything that is cached.
}
