% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{set_current_database}
\alias{set_current_database}
\title{Set The Current Database}
\usage{
set_current_database(sc, name)
}
\arguments{
\item{sc}{A \code{spark_connection}.}

\item{name}{\code{character(1)}. The name of the database to set the current
database to.}
}
\description{
Sets the current default database in this session.
}
\examples{
\dontrun{
sc <- sparklyr::spark_connect(master = "local")
set_current_database(sc = sc, name = "new_db")
}

}
\seealso{
\code{\link[=current_database]{current_database()}}, \code{\link[=database_exists]{database_exists()}}, \code{\link[=list_databases]{list_databases()}}
}
